fix: skip if already symlinked

+++ container_warmup.sh
@@ -11,19 +11,25 @@ if [ -r ./serverfiles_config/arkserver.cfg -a -n "$(ls ./serverfiles_saved/*.ark
   ln -s ../../../serverfiles_saved/SavedArks ./serverfiles/ShooterGame/Saved/SavedArks
   # move Logs
   mkdir ./serverfiles_saved/Logs
-  mv ./serverfiles/ShooterGame/Saved/Logs/* ./serverfiles_saved/Logs/
-  rmdir ./serverfiles/ShooterGame/Saved/Logs
-  ln -s ../../../serverfiles_saved/Logs ./serverfiles/ShooterGame/Saved/Logs
+  if [ ! -h ./serverfiles/ShooterGame/Saved/Logs ]; then
+    mv ./serverfiles/ShooterGame/Saved/Logs/* ./serverfiles_saved/Logs/
+    rmdir ./serverfiles/ShooterGame/Saved/Logs
+    ln -s ../../../serverfiles_saved/Logs ./serverfiles/ShooterGame/Saved/Logs
+  fi
   # move SaveGames
   mkdir ./serverfiles_saved/SaveGames
-  mv ./serverfiles/ShooterGame/Saved/SaveGames/* ./serverfiles_saved/SaveGames/
-  rmdir ./serverfiles/ShooterGame/Saved/SaveGames
-  ln -s ../../../serverfiles_saved/SaveGames ./serverfiles/ShooterGame/Saved/SaveGames
+  if [ ! -h ./serverfiles/ShooterGame/Saved/SaveGames ]; then
+    mv ./serverfiles/ShooterGame/Saved/SaveGames/* ./serverfiles_saved/SaveGames/
+    rmdir ./serverfiles/ShooterGame/Saved/SaveGames
+    ln -s ../../../serverfiles_saved/SaveGames ./serverfiles/ShooterGame/Saved/SaveGames
+  fi
   # move config
   mkdir ./serverfiles_saved/Config
   mv ./serverfiles_config/* ./serverfiles_saved/Config/
-  rm ./serverfiles/ShooterGame/Saved/Config
-  ln -s ../../../serverfiles_saved/Config ./serverfiles/ShooterGame/Saved/Config
+  if [ ! -h ./serverfiles/ShooterGame/Saved/Config ]; then
+    rm ./serverfiles/ShooterGame/Saved/Config
+    ln -s ../../../serverfiles_saved/Config ./serverfiles/ShooterGame/Saved/Config
+  fi
 
   echo "--> Migration complete: you no longer need to mount 'serverfiles_config'"
 fi
